refactor(transfer): split receiver/tests/file_list and errors_and_timeouts
Decomposes two oversized test files into directory modules so each
sub-file stays under the 650-line cap and groups one concern per file.

file_list.rs (1648 lines) -> file_list/{mod, wire_attrs, id_lists,
incremental_receiver, incremental_directories, filter_chain,
proto_io_error, ndx_convert, delete_pipeline_hook}.

errors_and_timeouts.rs (832 lines) -> errors_and_timeouts/{mod,
legacy_goodbye_tests, input_multiplex_tests, sanitize_file_list,
daemon_filter_tests}.

No public API change. Largest new file is incremental_receiver.rs at
430 lines.
